Plot alluvial/sankey diagram of transport mode changes

Takes two trips_table (from readTripsTable), and collects changes between transport mode distribution of these tables to make alluvial diagram from this data

Usage

plotModalShiftSankey(
  tripsTable1,
  tripsTable2,
  show.onlyChanges = FALSE,
  unite.columns = character(0),
  united.name = "united",
  dump.output.to = matsimDumpOutputDirectory
)

Arguments

tripsTable1

tible of trips_output (from readTripsTable())

tripsTable2

tible of trips_output (from readTripsTable())

show.onlyChanges

boolean, if it is set to TRUE => sankey diagram only contains changes on axes

unite.columns

vector of character string, changes name of all transport modes in the tibble copy to united.name = "united" that matches PATTERNS given in unite.columns

united.name

if columns were united, you can specify name for the resulting column in plot

dump.output.to

folder that saves and configures yaml for simwrapper. folder where png of plot is stored

Value

Alluvial diagram that represents changes in transport mode distribution of trip tables

Details

Function calculates number of each transport mode used in first and second table, and draws plot that represent how distribution of transport mode has changed (f. e. what part of concrete trasport mode changed to another) Using parameter unite.columns transport modes that match PATTERN in unite.columns can be united in 1 transport mode type (by default united.name is "united") Using parameter show.onlyChanges