Appends an additional column with the trip type to output_trips

Using a shape_file of the project area an additional column is created, categorizing all trips into the following categories. inside: trips that start and end inside the given shape originating: trips that start inside the shape and end outside of the shape destinating: trips that end inside the shape and start outside of the shape outside: trips that start and end outside of the shape

Usage

process_append_spatialcat(trips_table, shape_table, crs)

Arguments

trips_table

tibble of output_trips (from read_output_trips())

shape_table

sf object(data.frame with geometries), can be created using st_read()

crs

numeric representation of the EPSG code or proj4string for the corresponding coordinate system of the trip coordinates, can be found in network file from output directory of MATSim simulation

Value

tibble, with additional column containing the trip type

Details

Please be aware that this process_filter_by_shape currently only works, when one geometry is loaded.)